requirements.txt
paho-mqtt==2.0.0
pillow==10.3.0
